functions/analysis/clip-embedder/requirements.txt
boto3
vastdb
pyarrow
Pillow
open_clip_torch
torch
